Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> (1 comment)
> 
> lgtm with a comment, thanks.

I don't have direct access to a Windows ARM64 machine, a co-worker told me the tests were failing, and we concluded it was due to x18.
We didn't dig in to find if it was because windows was clobbering x18 or if us clobbering x18 was causing issues for windows.

Patch-set: 1
Reviewer: Gerrit User 1334495 <1334495@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>