Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hey Georg, this addition for API calls turned out to be really small and (mostly) orthogonal to the other CL, so PTAL, thanks!

That's a good start, but we'll have to do the same work here as in the other CL, i.e. we'll have to call ProcessReceiverMapForApiCall somewhere so that the call reducer will be able to do its job.

Patch-set: 1
Reviewer: Gerrit User 1119597 <1119597@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>