[ic] Load maps and handlers only once in the IC constructor.
Traditionally we had loaded the maps and handlers several times from the
FeedbackVector (via the nexus abstraction), which was mostly for historic
reasons. This changes the behavior of the IC class to only load the maps
and handlers from the FeedbackVector once in the beginning and then uses
them until the end.

R=ishell@chromium.org

Bug: v8:9183
Change-Id: I9319f711f8847d40573e18ff7315bc8034adbb04