Update patch set 2
Patch Set 2: Code-Review+1

> Patch Set 2:
> 
> Hi Michael,
> 
> This is a tiny change, but I'd appreciate a sanity check from an expert on allocation mementos. Do you remember why AllocationMemento::IsValid checks allocation_site()->IsAllocationSite()?
> 
> Thanks!

Hi Seth, thanks for the CL. LGTM. I remember there was a lot of caution in AllocationMementos, because they are tested for existence at a structural place (right behind an object) but they may not actually be there. Then, if one is found, there is the worry that the AllocationSite pointed to may be in a "zombie" state. However, looking at the codebase now, I think we can eliminate this verification that the AllocationSite is an AllocationSite.

Patch-set: 2
Reviewer: Gerrit User 1118564 <1118564@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1