Update patch set 1
Patch Set 1:

I probably triggered this renaming. I was not aware that a discussion hat lead to the name builtin pointer, and found it confusing, since it is a Smi expressing the index. I could be convinced to leave things as they are.

Patch-set: 1
Reviewer: Gerrit User 1119600 <1119600@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>