Add Base64 VLQ decoding support
This CL adds the support for Base64 VLQ decoding, which is the base of
parsing source map files (Version 3). With this support, the mappings of
C/C++ source code and WASM bytecode could be built in V8 engine. The
newly-added function is called VLQBase64Decode, which accepts two

argument: the Base64-encoded VLQ string and the position of first
character to be decoded. Upon its return, the position is updated with
the next start position. The unittest of this support is also added in
this CL.
Change-Id: If0f32972ecd7488844478a7b93a0f10cc38b6a5d