Update patch set 1
Patch Set 1: Code-Review+1

> Patch Set 1: Code-Review+1
> 
> Sounds goot to me. Please wait for Jakob's review as well.

So on the macro asm side this sgtm as well, but then the confusion may just be shifted to code-generator where kArchCallBuiltinPointer would now be implemented as CallBuiltinByIndex. 

What I like about the current naming is that all related code uses the same terms and can be found by a simple `git grep CallBuiltinPointer`. Note this is not just Torque, but threads through from CSA, the compiler, and finally into masm. We'd lose that with the rename.

As one more suggestion, wdyt about a clarifying comment next to CallBuiltinPointer that explains what a 'builtin pointer' is?

Given all this, if you're still convinced a rename is the way to go then I guess I'm okay with that.. So +1, although I wouldn't be unhappy either if we were to just add a comment instead.

Patch-set: 1
Reviewer: Gerrit User 1162439 <1162439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1