Update patch set 3
Patch Set 3:

Hi,

given that the native context's scopeinfo is a dummy (empty) and we don't care about other scopeinfos, these changes feel like overkill. Or are there other uses that you have in mind?

For the promise stuff alone, I think all we need is to change NativeContextRef/-Data to provide the (empty) ScopeInfoRef and serialize it either in the constructor or in the SerializeStandardObjects phase, i.e. no changes to ContextRef/-Data and no need for anything optional.

Patch-set: 3