Update patch set 2
Patch Set 2: Code-Review+1

As discussed on the bug, I think this is a good solution to the regression, but the long-term solution may rather be fixing the preparse data. Once that is done, maybe this CL can be reverted (or, maybe the risk+complexity isn't worth the memory savings, in which case we can still keep it).

Patch-set: 2
Reviewer: Gerrit User 1175774 <1175774@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1