WIP: Add postmortem debugging helper library
This is an incomplete prototype, shared to provide a sketch of my
current ideas for implementation of the "JavaScript object
investigation" section of this design document:
https://docs.google.com/document/d/1evHnb1uLlSbvHAAsmOXyc25x3uh1DjgNa8u1RHvwVhk/edit#

The primary library entry point is the function GetObjectProperties:
https://crrev.com/c/v8/v8/+/1628012/2/tools/debug_helper/debug-helper.h#161
This function takes the object pointer to inspect, a callback function
for reading memory, and some other optional data that might be helpful,
and returns a struct which contains both a brief string description of
the object and an array of properties the object contains.

The test file demonstrates a few simple cases of fetching properties:
https://crrev.com/c/v8/v8/+/1628012/2/test/cctest/test-debug-helper.cc

This change currently uses plain C structs on the library boundary, to
make linking/embedding easier from debugger extensions built with
alternative toolchains. The library uses proper ownership semantics and
standard collections everywhere up to the boundary function where it
must release and return a raw pointer. I'm not sure this is worth the
trouble though, since V8 embedders are already accustomed to linking
against a C++ API.

Bug: v8:9376

Change-Id: Iebf3cc2dea3133c7811bcefcdf38d9458b02fded