Move feedback for BigIntTooBig from deoptimizer to interpreter
Feedback shall not be updated by the deoptimizer. Although this
mechanism exists, it shall not be used if possible. This CL changes
how V8 learns from BigInt deopts: Previously we updated feedback on
the BinaryOperations in the deoptimizer, now we let the interpreter
widen the feedback type from BigInt to Any after the deopt has occurred.

Bug: v8:9407
Change-Id: I92e5e733085b433fd8ab452674d02404b81b2796