Update patch set 3
Patch Set 3:

Here is a simpler repro:

 const a = {toString: () => {
   console.log("print arguments", print.arguments);
 }};

 function g(x) {
   print(x);
 }

 %PrepareFunctionForOptimization(g);
 g(a);
 g(a);
 %OptimizeFunctionOnNextCall(g);
 g(a);

In general, this means that the fake-frame arguments we use are observable from JS. This means that we have to think more about this; for example, HTMLElement.arguments can also be read. I can't think of a way to trigger inlining of HTMLElement at the moment, but there might be more functions that allow reading their .arguments.

Patch-set: 3
Reviewer: Gerrit User 1119600 <1119600@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>