Update patch set 7
Patch Set 7:

Changed the regression test to Sigurds repro. Thanks for your help on this :-)

I propose we land the fix as-is for now to fix the crash and implement proper {argument} handling for API functions in a later CL (will be a non-trivial change).

Patch-set: 7
Reviewer: Gerrit User 1323977 <1323977@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0