[ptr-compr][cleanup][turbofan] Narrow Mach Rep scope in WriteBarrierKindFor
If we are on pointer compression, we are only going to be storing
Compressed values. If we are not, we are only going to be storing Tagged
values. The other Machine Representation is not needed.

Cq-Include-Trybots: luci.v8.try:v8_linux64_pointer_compression_rel_ng
Cq-Include-Trybots: luci.v8.try:v8_linux64_arm64_pointer_compression_rel_ng
Bug: v8:7703, v8:9396
Change-Id: I6c821b629635ce91eae3589508c07a0ae4e5606a