[cpu-profiler] Fix a failing DCHECK
I noticed that this DCHECK was failing when I attempted to collect a CPU
profile on JetStream 2. (I also noticed that I had DCHECKs enabled,
which means that profiling is not a particularly valuable activity.) We
can allow multiple equivalent values in a row and still assert that the
vector is sorted, as the nearby comment states. I haven't looked in
depth at why we have multiple equivalent values or whether it's likely
to cause problems elsewhere.

Change-Id: I9daf029eb995cd16a79c1bdc386bfdc69f33f93c