Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Thanks! Out of curiosity, what is experiment_percentage?

The bot runs with this percentage on CLs in CQ, but does not block. You can show the experimental bots with a button in gerrit. E.g. it also ran on the CL that caused a breakage yesterday and we can see that it would have prevented it.

Do you also +1?

Patch-set: 1
Reviewer: Gerrit User 1115956 <1115956@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0