Update patch set 3
Patch Set 3:

It's not that we didn't want to do the optimization, but that we were afraid it's unsound. I don't see how pointer compression would be different from 32bit builds in this respect. 
I'm not sure, but I think the problem was that we need to invalidate load information state for the aliased second tagged-sized chunk of a big field.

Patch-set: 3