Update patch set 4
Patch Set 4:

(4 comments)

Thanks for working on this fix. After seeing this code for the first time, and your changes  that made it even tougher to read (I didn't finish actually), I'm thinking that it would be much nicer to have this as a Torque builtin. Would you mind trying that? It would mean converting the existing builtin to a Torque javascript builtin.

If this turns out to be too slow, we can turn that into a Torque non-javascript builtin that will be faster to call into but will have to be restricted to a fixed number of arguments.

Patch-set: 4
CC: Gerrit User 1180419 <1180419@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1119597 <1119597@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>