Add ApiNatives::InstantiateFunction that explicity takes native context
This CL changes {CreateApiFunction} to take an explicit native context
for the newly created JSFunction. This is a refactoring in preparation
for a bugfix. AccessorPairs can be instantiatd lazily. At the time of
lazy instantiation, the current context does not necessarliy match
the creation context of the holder of an AccessorPair.

R=bmeurer@chromium.org, verwaest@chromium.org

Bug: chromium:986063, chromium:989909
Change-Id: Idea4b5052f2baff5c3d916f5ab8ed5017b60699b