Record native context of AccessorPairs when installing break trampolines
This CL changes the way AccessorPairs are collected for instantiation.
Instead of walking the heap and looking at AccessorPairs directly, we
look at all JSObjects and collect AccessorPair via each objects
descriptor array. This way, we can associate the correct native
context with each collected AccessorPair.

The current native context is not always the correct context to instantiate
the getter and setter JSFunctions for an AccessorPair.

Bug: chromium:986063
Change-Id: I124a0802f4938b95f1ad75efc65eb05b66bcfc67