Update patch set 9
Patch Set 9:

(1 comment)

Thanks a lot for these thoughts and comments!

> Have there been thoughts about this from others?

I discussed the approach with Georg.

> - The mismatch with the runtime type hierarchy. People that don't know Torque will not know to use JSAny when writing new Torque code. And V8 devs reading Torque code will likely have no clue what 'JSAny' is. I guess that's an okay-ish trade-off since JSAny gives stronger guarantees in Torque, but I know folks will run into this.

That's true, and I don't really have an idea how to avoid this. To some extend the type system will force people to use JSAny correctly, but still, it adds to the learning curve.

> - JSAny in general does not obviously mean 'any JS-exposed Object' to me.

I'll add a comment to the definition of JSAny to convey the idea. "JSAny" is inspired by Typescript's "any" type, with which it should actually coincide. We should use something short, since it will be one of the most frequent types in Torque.

> - The JSAny/AnyName combination could be slightly confusing.

Yes, this is a temporary solution. I plan to add a language feature such that AnyName becomes unnecessary and we can use Name directly.

> - JSAnyNotSmi: This is basically JSAny for HeapObjects, yet that's not clear from the name. Again, probably okay but there's confusion potential.

True, but I don't have a better idea, JSAnyHeapObject and HeapObjectJSAny are even more confusing.

> Just out of curiosity, would annotations (on types) help here to avoid building a small parallel type hierarchy?

Interesting idea, but how would this work? Keeping the CSA-version of types as part of the name is potentially very verbose. The type hierarchy is not really parallel, just more refined. JSAny is still a subtype of the Torque Object type, so in a way, the current solution should be the same as type annotations except for not explicitly mentioning the CSA version.

> Another idea without addtl types: what about just implicitly checking boundaries to JS? I.e. Call(), incoming parameters and return values for JS builtins.

That would help too for sure. Actually, we currently don't even have that, since the transitions between Torque and CSA are completely unchecked, but we should address that. Something valuable from adding JSAny we wouldn't get with this approach is that code now documents what is a JavaScript value and what isn't, and when we're using TheHole.

Patch-set: 9
Reviewer: Gerrit User 1180419 <1180419@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>