Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> > Patch Set 9:
> > 
> > (1 comment)
> > 
> > Thanks a lot for these thoughts and comments!
> > 
> > > Have there been thoughts about this from others?
> > 
> > I discussed the approach with Georg.
> > 
> > > - The mismatch with the runtime type hierarchy. People that don't know Torque will not know to use JSAny when writing new Torque code. And V8 devs reading Torque code will likely have no clue what 'JSAny' is. I guess that's an okay-ish trade-off since JSAny gives stronger guarantees in Torque, but I know folks will run into this.
> > 
> > That's true, and I don't really have an idea how to avoid this. To some extend the type system will force people to use JSAny correctly, but still, it adds to the learning curve.
> > 
> 
> Could this be avoided by adding a JSAny to the runtime hierarchy as well? I'd expect torque types to match 1:1 and would definitely be confused (at least initially)

It's not possible to add it to the existing hierarchy as is, because it very much is a union type. In theory, we could try to emulate it in the runtime with a special class and appropriate C++ implicit conversions. That could either be a one-off for JSAny with implicit conversions from Smi, HeapNumber, Name, JSReceiver and some Oddball subclasses that would need to be created, or we could add a general kind of union type mechanism to the runtime. I once attempted to do that for the CSA type system: https://cs.chromium.org/chromium/src/v8/src/compiler/code-assembler.h?sq=package:chromium&dr&q=UnionT
This code is somewhat abandoned, but it's quite possible to do the same in the runtime. However, it will be quite template-metaprogramming style.

In general, I don't think we'll be able to keep this up for all Torque types that are more fine-grained than their runtime equivalent. For example, transient types shouldn't be added to the runtime I think since we can't have the appropriate checking there. And restricting Torque to only have the runtime types is very much in conflict with the goals we have for Torque.

Patch-set: 10
Reviewer: Gerrit User 1180419 <1180419@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0